backend/requirements_render.txt
# Core FastAPI and server
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Document processing
PyMuPDF==1.25.2
python-docx==1.1.2
pillow==11.1.0

# LangChain framework
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-pinecone==0.2.0
langchain-community==0.3.14
langchain-text-splitters==0.3.4

# Google AI
google-generativeai==0.8.5

# Pinecone vector database
pinecone==5.4.0

# Environment and configuration
python-dotenv==1.0.1

# Data processing
numpy==2.2.0
pydantic==2.11.7

# Additional dependencies for production
gunicorn==21.2.0
aiofiles==24.1.0
